requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
sqlparse
tzdata
coreapi